requirements.txt
numpy
opencv-python==4.1.2.30
google-cloud-storage
